Crate subcoin_runtime

Subcoin runtime is a minimalistic Substrate runtime consisting of frame-system and pallet-bitcoin. It does not implement all the typical runtime APIs the normal runtimes would do as many of them does not make sense in Subcoin.

Modules

api

interface

Some re-exports that the node side code needs to know. Some are useful in this context as well.

Structs

BlockWeights

Subcoin block weights.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

Version

Enums

OriginCaller

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

Constants

RUNTIME_API_VERSIONS

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

WASM_BINARY_PATH

Functions

native_version

The version information used to identify this runtime when compiled natively.

Type Aliases

Address

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Bitcoin

Header

System

SystemConfig

UncheckedExtrinsic