subcoin_runtime

Type Alias UncheckedExtrinsic

pub type UncheckedExtrinsic = UncheckedExtrinsic<Address, RuntimeCall, MultiSignature, (CheckNonZeroSender<Runtime>, CheckSpecVersion<Runtime>, CheckGenesis<Runtime>)>;

Aliased Type

pub struct UncheckedExtrinsic {
    pub preamble: Preamble<MultiAddress<AccountId32, ()>, MultiSignature, (CheckNonZeroSender<Runtime>, CheckSpecVersion<Runtime>, CheckGenesis<Runtime>)>,
    pub function: RuntimeCall,
    pub encoded_call: Option<Vec<u8>>,
}

Fields
preamble: Preamble<MultiAddress<AccountId32, ()>, MultiSignature, (CheckNonZeroSender<Runtime>, CheckSpecVersion<Runtime>, CheckGenesis<Runtime>)>

Information regarding the type of extrinsic this is (inherent or transaction) as well as associated extension (Extension) data if it’s a transaction and a possible signature.

function: RuntimeCall

The function that should be called.

encoded_call: Option<Vec<u8>>

Stores the raw encoded call.

This is mainly interesting if this extrinsic was created by decoding it from bytes. In this case this field should be set to Some holding the original bytes used to decode the [Self::function]. This is done to protect against decode implementations of Call that are not bijective (encodes to the exact same bytes it was encoded from). If this field is set, it is being used when re-encoding this transaction.